바우처 관리 어플리케이션 단위테스트 작성, 코드 리팩토링
- 바우처 관련 모든 클래스에 대한 단위테스트 작성
- VoucherController에서 CommandLine에 대한 의존성을 주입받도록 변경(CommandLineView를 빈으로 생성 및 생성자 주입)
- VoucherRepository 파일 경로를 Property로 설정

+++ kdt-spring-order/src/main/java/org/prgrms/kdt/voucher/domain/PercentDiscountVoucher.java
@@ -29,7 +29,7 @@ public class PercentDiscountVoucher implements Voucher {
 
     @Override
     public long discount(long beforeDiscount) {
-        return beforeDiscount * (percent/100);
+        return (long)(beforeDiscount * (percent/100.0));
     }
 
     @Override
